# oracleforge/landing/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MAXIA Oracle",
    "version": "0.1.9",
    "description": "Multi-source price feed for AI agents. Data feed only. Not investment advice. No custody. No KYC.",
    "contact": {
      "name": "MAXIA",
      "email": "ceo@maxiaworld.app",
      "url": "https://oracle.maxiaworld.app"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "paths": {
...
    "/api/alerts": {
      "post": {
        "tags": [
          "alerts"
        ],
        "summary": "Create Alert Route",
        "description": "Create a price alert. One-shot: triggers once then deactivates.",
        "operationId": "create_alert_route_api_alerts_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "List Alerts Route",
        "description": "List all alerts for the authenticated key.",
        "operationId": "list_alerts_route_api_alerts_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/alerts/{alert_id}": {
      "delete": {
        "tags": [
          "alerts"
        ],
        "summary": "Delete Alert Route",
        "description": "Delete an alert by id.",
        "operationId": "delete_alert_route_api_alerts__alert_id__delete",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Alert Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prices/stream": {
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "Stream Prices",
        "description": "SSE stream of live prices, polled every ~3 seconds.\n\nDoes not consume daily quota. Auto-closes after 1 hour.",
        "operationId": "stream_prices_api_prices_stream_get",
        "parameters": [
          {
            "name": "symbols",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Comma-separated symbols (max 10), e.g. BTC,ETH,SOL",
              "title": "Symbols"
            },
            "description": "Comma-separated symbols (max 10), e.g. BTC,ETH,SOL"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AlertCreateRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "maxLength": 10,
            "minLength": 1,
            "title": "Symbol"
          },
          "condition": {
            "type": "string",
            "enum": [
              "above",
              "below"
            ],
            "title": "Condition"
          },
          "threshold": {
            "type": "number",
            "maximum": 1000000000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Threshold"
          },
          "callback_url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 10,
            "title": "Callback Url"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "condition",
          "threshold",
          "callback_url"
        ],
        "title": "AlertCreateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "https://oracle.maxiaworld.app",
      "description": "Production"
    }
  ]
}